moving collect unbound variables method to utils
