+++ src/passes/externalArgumentModifier/structModifier.ts
@@ -15,9 +15,9 @@ import {
 import { AST } from '../../ast/ast';
 import { ASTMapper } from '../../ast/mapper';
 import { cloneASTNode } from '../../utils/cloning';
+import { collectUnboundVariables } from '../../utils/functionGeneration';
 import { createIdentifier } from '../../utils/nodeTemplates';
 import { isExternallyVisible } from '../../utils/utils';
-import { collectUnboundVariables } from './../loopFunctionaliser/utils';
 
 export class StructModifier extends ASTMapper {
   /*
