+++ tests/behaviour/contracts/loops/loops.sol
@@ -12,78 +12,78 @@ contract WARP {
     return x;
   }
 
-  // function whileLoop(uint8 target) pure public returns (uint8) {
-  //   uint8 x = 0;
-  //   while(x != target) {
-  //     x++;
-  //   }
-  //   return x;
-  // }
+  function whileLoop(uint8 target) public pure returns (uint8) {
+    uint8 x = 0;
+    while (x != target) {
+      x++;
+    }
+    return x;
+  }
 
-  // function innerReturn(uint8 target) pure public {
-  //   uint8 x = 0;
-  //   while(x != target) {
-  //     return;
-  //   }
-  //   revert();
-  // }
+  function innerReturn(uint8 target) public pure {
+    uint8 x = 0;
+    while (x != target) {
+      return;
+    }
+    revert();
+  }
 
-  // function breaks(uint8 target) pure public returns (uint8) {
-  //   uint8 x = 0;
-  //   while(x != target * 2) {
-  //     if (x == target) break;
-  //     ++x;
-  //   }
-  //   return x;
-  // }
+  function breaks(uint8 target) public pure returns (uint8) {
+    uint8 x = 0;
+    while (x != target * 2) {
+      if (x == target) break;
+      ++x;
+    }
+    return x;
+  }
 
-  // function continues(uint8 count) pure public returns (uint8) {
-  //   uint8 x = 3;
-  //   for(uint8 i = 0; i < count; ++i) {
-  //     if (i == 2) continue;
-  //     x *= 2;
-  //   }
-  //   return x;
-  // }
+  function continues(uint8 count) public pure returns (uint8) {
+    uint8 x = 3;
+    for (uint8 i = 0; i < count; ++i) {
+      if (i == 2) continue;
+      x *= 2;
+    }
+    return x;
+  }
 
-  // function doWhile(uint8 init, uint8 target) pure public returns (uint8) {
-  //   do {
-  //     init++;
-  //   } while(init <= target);
+  function doWhile(uint8 init, uint8 target) public pure returns (uint8) {
+    do {
+      init++;
+    } while (init <= target);
 
-  //   return init;
-  // }
+    return init;
+  }
 
-  // function doWhile_break(uint8 init, uint8 target) pure public returns (uint8) {
-  //   do {
-  //     if (init == target) break;
-  //     init++;
-  //   } while(init <= target * 2);
+  function doWhile_break(uint8 init, uint8 target) public pure returns (uint8) {
+    do {
+      if (init == target) break;
+      init++;
+    } while (init <= target * 2);
 
-  //   return init;
-  // }
+    return init;
+  }
 
-  // function doWhile_continue(uint8 target) pure public returns (uint8) {
-  //   uint8 x = 0;
+  function doWhile_continue(uint8 target) public pure returns (uint8) {
+    uint8 x = 0;
 
-  //   do {
-  //     x++;
-  //     if (x == target) continue;
-  //   } while(x != target);
+    do {
+      x++;
+      if (x == target) continue;
+    } while (x != target);
 
-  //   return x;
-  // }
+    return x;
+  }
 
-  // function doWhile_return(uint8 target) pure public returns (uint8) {
-  //   uint8 x = 0;
+  function doWhile_return(uint8 target) public pure returns (uint8) {
+    uint8 x = 0;
 
-  //   do {
-  //     x++;
-  //     if (x == 2) {
-  //       return x;
-  //     }
-  //   } while(x <= target);
+    do {
+      x++;
+      if (x == 2) {
+        return x;
+      }
+    } while (x <= target);
 
-  //   return x;
-  // }
+    return x;
+  }
 }
